ci: consolidate workflows, add caching, fix compiler warning
- Merge build.yml into ci.yml (removes duplication)

- Add CMake caching to all platforms (faster builds)

- Fix signed/unsigned comparison warning in AestraThreading.h

- Add pre-commit hook for platform abstraction checks

- Format check runs but doesn't block (continue-on-error)

+++ AestraCore/include/AestraThreading.h
@@ -271,7 +271,7 @@ private:
     void workerLoop(uint32_t threadIdx) {
         (void)threadIdx;
         while (!m_stop) {
-            m_signal.wait([this] { return m_stop || m_taskCounter.load(std::memory_order_acquire) < m_taskCount; });
+            m_signal.wait([this] { return m_stop || static_cast<uint32_t>(m_taskCounter.load(std::memory_order_acquire)) < m_taskCount; });
 
             if (m_stop)
                 return;
